% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{transform_inputs}
\alias{transform_inputs}
\title{Inputted data expected for each golf round:
Date
Course
Rating
Slope
Par
Duration
Transport
Gross score
Fairways in Reg
Fairways to Hit
Greens in Reg
Putts}
\usage{
transform_inputs(input_data)
}
\description{
Methods to compute the following:
Over/Under (Gross)
Course Handicap
Net score
Over/Under (Net)
Resultant Handicap Index (Adjusts based on number of prior samples)
Handicap Differential (Used for Course Handicap and Handicap Index)
}
\details{
Other code (scripts/shiny app) will handle:
Display of per round data
Display of summary data
Trending of performance over time, pivoting on various features
}
